pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "foxess-inverter"
version = "0.1.0"
description = "Python client for FoxESS Cloud API - inverter management"
requires-python = ">=3.11"
dependencies = ["requests>=2.28", "wasmtime>=29.0.0"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pre-commit>=3.0",
    "mypy>=1.15",
    "ruff>=0.11",
    "types-requests>=2.28",
    "homeassistant",
    "voluptuous",
    "PyJWT",
    "pytest-playwright",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: end-to-end tests requiring podman and HA container (run with pytest e2e/ -m slow)",
]

[tool.mypy]
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "custom_components.foxess_control"
# websocket_api attrs resolve in full HA but not in pre-commit's isolated env.
disable_error_code = ["unused-ignore"]

[[tool.mypy.overrides]]
module = "custom_components.foxess_control.config_flow"
# HA types resolve as Any in pre-commit's isolated mypy env,
# causing false positives on ConfigFlow/OptionsFlow subclassing.
disable_error_code = ["misc", "call-arg"]

[[tool.mypy.overrides]]
module = ["tests.test_init", "tests.test_entity_mode"]
# Tests import private aliases (_calculate_charge_power etc.) from
# foxess_control.__init__ which are not explicit re-exports.
disable_error_code = ["attr-defined"]

[[tool.mypy.overrides]]
module = "tests.test_config_flow"
# Assigning MagicMock to HA flow methods for testing.
disable_error_code = ["method-assign"]

[tool.ruff]
extend-exclude = ["*.json", "*.md"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH"]
